updated RemoteProductServiceTest.java

+++ product-microservice/src/main/java/com/metao/book/retails/domain/ProductEntity.java
@@ -8,6 +8,23 @@ import com.metao.book.shared.domain.base.AbstractAggregateRoot;
 import com.metao.book.shared.domain.base.ConcurrencySafeDomainObject;
 import com.metao.book.shared.domain.financial.Currency;
 import com.metao.book.shared.domain.financial.Money;
+import java.math.BigDecimal;
+import java.util.HashSet;
+import java.util.Set;
+import java.util.stream.Collectors;
+import javax.persistence.CascadeType;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.EnumType;
+import javax.persistence.Enumerated;
+import javax.persistence.FetchType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.JoinTable;
+import javax.persistence.ManyToMany;
+import javax.persistence.Table;
+import javax.persistence.Version;
+import javax.validation.Valid;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -16,12 +33,6 @@ import org.hibernate.annotations.Cache;
 import org.hibernate.annotations.CacheConcurrencyStrategy;
 import org.hibernate.validator.constraints.Length;
 import org.springframework.lang.NonNull;
-
-import javax.persistence.*;
-import javax.validation.Valid;
-import java.util.HashSet;
-import java.util.Set;
-import java.util.stream.Collectors;
 
 @Entity
 @Getter
@@ -48,7 +59,7 @@ public class ProductEntity extends AbstractAggregateRoot<ProductId> implements C
     private Image image;
 
     @Column(name = "price_value", nullable = false)
-    private Double priceValue;
+    private BigDecimal priceValue;
 
     @Valid
     @Column(name = "price_currency", nullable = false)
